libcamera: v4l2_controls: Fix usage of strerror()
On failure, the return code from V4L2Device::ioctl() is the negative error
code set by the failed ::ioctl() system call. When the return code of
V4L2Device::ioctl() is provided to strerror() it has to be negated again
to obtain the positive error code. Fix a few wrong usages of the return
code which provided to the strerror() function a negative error code.

Reviewed-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Signed-off-by: Jacopo Mondi <jacopo@jmondi.org>