test: ipa: Fix path to VIMC IPA
When changing the VIMC IPA source file structure the path in the
ipa_module_test which needs access to the .so file was not updated, fix
that.

The omission was hard to spot as it requires a clean build of the
project as the old .so file is still in the build tree and the test
passes.

Fixes: a25533089bda04da ("ipa: Move vimc to a subdirectory")
Signed-off-by: Niklas Söderlund <niklas.soderlund@ragnatech.se>
Reviewed-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Reviewed-by: Kieran Bingham <kieran.bingham@ideasonboard.com>