libcamera: controls: Add AWB related controls
AwbMode is a new enum type to specify operating mode of the AWB
algorithm. All modes may not be supported by all platforms.

ColourGains is a new float array type used to specify manual red
and blue (in that order) colour channel gains when AWB is disabled.

ColourTemperature is a new control to return the current estimate of the
colour temperature.

Signed-off-by: Naushir Patuck <naush@raspberrypi.com>
Reviewed-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>