libcamera: ipu3: Remove id from camera names
The id in the camera name is confusing and is of little use for users.
Camera names are not (yet) required to be unique and appending which
numerical CIO2 unit the sensor is attached to is just as good as
depending on the i2c bus information already present in the entity name.

Before this change,

    $ cam -l
    Available cameras:
    1: ov13858 2-0010 0
    2: ov5670 4-0036 1

After this change,

    $ cam -l
    Available cameras:
    1: ov13858 2-0010
    2: ov5670 4-0036

Signed-off-by: Niklas Söderlund <niklas.soderlund@ragnatech.se>
Reviewed-by: Jacopo Mondi <jacopo@jmondi.org>
Reviewed-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>